Feat: Increment hits by 1, and get current hits
Write two functions: First function, hits, which adds 1 to the value of hits every time it's called to represent when the battleship has been 'hit' in the game. Second function getHits, returns the current value for hits which can be used to track how many hits a ship has received which can be utilized to determine when the ship has sunk.