Feat: UI Alerts user not to skip their turn!
Pass/Done btns were disabled, so that the user could not unintentionally skip their turn, by clicking the button before submiting their attack. However, the UI/UX did not make it clear what the user should do, so toggle function updated to toggle event listeners on/off, and create an alert for a "disabled" status that tells the user what to do.