Style: Recalculate the placement-screen padding
Ship placement screen contains the gameboard grid and ship buttons to the side, but visually the game board should sit in the center of the screen, which does not occur when using flex-box center due to the ship buttons. Since, a custom padding-left is used to offset this issue, but a y-axis was added, this value was udpated, and additional comments provided to explain.