Feat: Remove redundant game start screen
When game is over, and user clicks play again, it is redundant to return to the start game screen consisting of a start button. It makes more sense to go to the select players screen. However the easiest way to reset the game is to just run location.load, which meant that it was easier to remove the the start screen all together and user flow instead begins on select players screen.